+++ examples/simple_tuned_model.jl
@@ -7,8 +7,7 @@ if ! ("ControlPlots" ∈ keys(Pkg.project().dependencies))
     using TestEnv; TestEnv.activate()
 end
 
-using SymbolicAWEModels, VortexStepMethod, KiteUtils, ControlPlots, Statistics
-using OrdinaryDiffEqCore
+using SymbolicAWEModels, ControlPlots
 
 set = Settings("system.yaml")
 sam = SymbolicAWEModel(set, "ram")
@@ -20,7 +19,7 @@ init!(tether_sam)
 
 simple_set = Settings("system.yaml")
 simple_sam = SymbolicAWEModel(simple_set, "simple_ram")
-init!(simple_sam; create_control_func=true)
+init!(simple_sam)
 
 sim_oscillate!(sam; total_time=1.0)
 SymbolicAWEModels.copy_to_simple!(sam, tether_sam, simple_sam)